CI: Test the compatibility between the development and release versions
At the moment, there is no constraint between the versions of the
OCaml/Unikraft packages (aka the ones currently using the version number
1.0.0) and the Unikraft ones (aka the ones currently using the version
number 0.18.0). So this adds the possibility to test in CI the
compatibility between various versions, in particular between packages
released to the opam-repository and packages in development.

CI: Reenable OCaml/Unikraft 1.0.0