Disable the work-tree lookup in `git apply`
As reported by Romain Calascibetta on other related projects such as
OCaml/Solo5, the use of `git apply` as a `patch` command can fail:

- `git apply` works as it was intended when it is not run in a git
  repository and when it is run at the root of the work tree; when it
  is launched in a sub-directory of a git repository, it’s behaviour is
  unclear: it seems just to skip the patches;
- `git apply` exits on success even if it skipped the patches, which
  makes it hard to diagnosticate the issue.

Note that launching `git apply` in the problematic case (a sub-directory
of git repository) happens when installing the opam package (aka the
sources are a tarball, not the source git repository) in an opam local
switch in git project.

Telling `git` that it is launched at the root of our project work tree
fixes the issue, even if the absence of failure exit code is rather
brittle. The `-v` verbose mode can help at least diagnosticate the issue
when the full log is available.